+++ helper-scripts/src/stackManagementTool/delete/deleteDevStacks.ts
