+++ helper-scripts/src/stackManagementTool/delete/deleteStacks/deleteStacks.ts
