+++ helper-scripts/src/stackManagementTool/delete/deleteStacks/prompts.ts
