+++ helper-scripts/src/stackManagementTool/protectedStacks/protectedStacks.ts
@@ -1,0 +1,9 @@
+export const protectedStacks = [
+  "mob-sts-mock",
+  "mob-sts-mock-tir",
+  "mob-sts-mock-pl",
+  "mob-async-backend",
+  "mob-async-backend-tir",
+  "mob-async-backend-pl",
+  "mob-async-backend-cf-dist",
+];
